% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_bound.R
\name{city_bound}
\alias{city_bound}
\title{Construct a shapefile containing city boundaries based on high-resolution population map.}
\usage{
city_bound(pop, threshold, dist, pop_sum)
}
\arguments{
\item{pop}{object of class raster containing gridded population data.}

\item{threshold}{object of class numeric; used as a threshold to define which areas should be considered as part of city.}

\item{dist}{object of class numeric; used as a threshold to be used as a radius/shell size of the burning procedure in cca().}

\item{pop_sum}{object of class numeric; this sets a threshold for the minimum population size of a geographical cluster to qualify as a urban agglomeration.}
}
\value{
`city_bound()` returns a sf object with polygon boundaries of urban agglomerations with a population size greater than pop_sum.
}
\description{
Construct a shapefile containing city boundaries based on high-resolution population map.
}
\examples{
\dontrun{
filename <- paste0("inst/extdata/zanzibar_ppp_2020_constrainted.tif")
pop <- raster(filename) \%>\%
raster::calc(function(x) ifelse(is.na(x), 0, x)) \%>\%
raster::aggregate(., fact=10, "sum")
threshold <- 1000
dist <- 1000
pop_sum <- 5000
pop <- raster::raster(filename)
city_bound(pop,threshold, dist, pop_sum)
}
}
