% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friction_matrix.R
\name{friction_matrix}
\alias{friction_matrix}
\title{Calculate distance matrix using your own specified friction map.}
\usage{
friction_matrix(x, y, friction)
}
\arguments{
\item{x}{a numeric vector of longitude and latitude (WGS84), an sf object of the origin point.}

\item{y}{a numeric vector of longitude and latitude (WGS84), an sf object of the origin point.}

\item{friction}{friction maps (available in GEOTIFF) which can be downloaded from https://malariaatlas.org/research-project/accessibility_to_cities/ or https://malariaatlas.org/research-project/accessibility-to-healthcare/.}
}
\value{
friction_matrix() returns a data.frame object with id_o (row ID of x), id_d (row ID of y), and dur (travel time between each element of x and that of y) in minutes.
}
\description{
Calculate distance matrix using your own specified friction map.
}
\examples{
tz <- read.csv("data/tz.csv") \%>\%  st_as_sf(coords = c('lng', 'lat'),crs=st_crs(4326))
friction <- raster("data/2015_friction_surface_v1.geotiff") ## this needs to be downloaded
friction_matrix(tz, tz, friction)
}
