% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm_matrix.R
\name{osrm_matrix}
\alias{osrm_matrix}
\title{Calculate distance matrix based on OpenStreetMap-Based Routing Service OSRM.}
\usage{
osrm_matrix(x, y, osrm.profile)
}
\arguments{
\item{x}{a numeric vector of longitude and latitude (WGS84), an sf object of the origin point.}

\item{y}{a numeric vector of longitude and latitude (WGS84), an sf object of the origin point.}

\item{osrm.profile}{the routing profile to use, e.g. "car", "bike" or "foot" (when using the routing.openstreetmap.de test server). "car" is the default.}
}
\value{
osrm_matrix() returns a data.frame object with id_o (row ID of x), id_d (row ID of y), and dur (travel time between each element of x and that of y) in minutes.
}
\description{
Calculate distance matrix based on OpenStreetMap-Based Routing Service OSRM.
}
\examples{
\dontrun{
tz <- read.csv("data/tz.csv") \%>\%  st_as_sf(coords = c('lng', 'lat'),crs=st_crs(4326))
osrm_matrix(tz, tz, "car")
}
}
