% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pts.R
\name{count_pts}
\alias{count_pts}
\title{Count the number of points by location.}
\usage{
count_pts(sf, pts, id)
}
\arguments{
\item{sf}{object of class sf containing administrative boundaries.}

\item{pts}{object of class sf containing points.}

\item{id}{location ID used to aggregate the number of points.}
}
\value{
`count_pts()` returns a data.frame object with id (location ID) and number of points by id.
}
\description{
Count the number of points by location.
}
\examples{
\dontrun{
count_pts(admin, tz, id)
}
}
